Move datasource error to repository package
