+++ internal/core/application/days.go
@@ -3,6 +3,7 @@ package application
 import (
 	"errors"
 	"github.com/danielcosme/curious-ape/internal/core/entity"
+	"github.com/danielcosme/curious-ape/internal/core/repository"
 	"github.com/danielcosme/curious-ape/internal/datasource"
 	"github.com/danielcosme/curious-ape/sdk/dates"
 	"time"
@@ -24,7 +25,7 @@ func (a *App) DaysGetAll() ([]*entity.Day, error) {
 
 func (a *App) DayGetByDate(date time.Time) (*entity.Day, error) {
 	d, err := a.db.Days.Get(entity.DayFilter{Date: date})
-	if err != nil && errors.Is(err, entity.ErrNotFound) {
+	if err != nil && errors.Is(err, repository.ErrNotFound) {
 		// if it does not exist, create new and return.
 		return a.DayCreate(&entity.Day{Date: date})
 	}
