+++ internal/datasource/sqlite/oauths.go
@@ -3,6 +3,7 @@ package sqlite
 import (
 	"fmt"
 	"github.com/danielcosme/curious-ape/internal/core/entity"
+	"github.com/danielcosme/curious-ape/internal/core/repository"
 	"github.com/jmoiron/sqlx"
 )
 
@@ -42,7 +43,7 @@ func (ds *Oauth2DataSource) Get(filter entity.Oauth2Filter) (*entity.Oauth2, err
 		q = fmt.Sprintf("%s %s", q, "WHERE provider = ?")
 		return o, catchErr(ds.DB.Get(o, q, filter.Provider))
 	}
-	return nil, entity.ErrNotFound
+	return nil, repository.ErrNotFound
 }
 
 func (ds *Oauth2DataSource) Find(filter entity.Oauth2Filter) ([]*entity.Oauth2, error) {
